Optimize 2015 day 4, fix day 15

+++ src/main/zig/2015/day15.zig
@@ -54,7 +54,7 @@ pub fn run(problem: *aoc.Problem) !aoc.Solution {
             }
         }
     }
-    return aoc.Solution{ .p1 = @intCast(usize, best), .p2 = @intCast(u16, best_500) };
+    return aoc.Solution{ .p1 = @intCast(usize, best), .p2 = @intCast(usize, best_500) };
 }
 
 fn int_from_tokens(tokens: *std.mem.TokenIterator, offset: usize) !i32 {
